+++ Products/CMFPlacefulWorkflow/DefaultWorkflowPolicy.py
@@ -198,7 +198,7 @@ class DefaultWorkflowPolicyDefinition(SimpleItemWithProperties):
         """
 
         cbt = self._chains_by_type
-        if isinstance(ob, type('')):
+        if isinstance(ob, basestring):
             pt = ob
         elif hasattr(aq_base(ob), '_getPortalTypeName'):
             pt = ob._getPortalTypeName()
@@ -213,7 +213,7 @@ class DefaultWorkflowPolicyDefinition(SimpleItemWithProperties):
             chain = cbt.get(pt, _MARKER)
 
         # Backwards compatibility: before chain was a string, not a list
-        if chain is not _MARKER and isinstance(chain, type('')):
+        if chain is not _MARKER and isinstance(chain, basestring):
             chain = map(lambda x: x.strip(), chain.split(','))
 
         Log.debug('Chain founded in policy %s', chain)
@@ -237,7 +237,7 @@ class DefaultWorkflowPolicyDefinition(SimpleItemWithProperties):
         """ Sets the default chain for this tool. """
         wftool = getToolByName(self, 'portal_workflow')
 
-        if isinstance(default_chain, type('')):
+        if isinstance(default_chain, basestring):
             default_chain = map(lambda x: x.strip(), default_chain.split(','))
         ids = []
         for wf_id in default_chain:
@@ -277,7 +277,7 @@ class DefaultWorkflowPolicyDefinition(SimpleItemWithProperties):
         if portal_type not in [pt.id for pt in self._listTypeInfo()]:
             raise ValueError("'%s' is not a valid portal type." % portal_type)
 
-        if isinstance(chain, type('')):
+        if isinstance(chain, basestring):
             chain = map(lambda x: x.strip(), chain.split(','))
 
         wftool = getToolByName(self, 'portal_workflow')